requirements.txt
google-generativeai
tenacity
google-patent-scraper
pypatent